requirements.txt
# Sentinel - Real-Time Financial Risk Agent
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Vector Store & Embeddings
chromadb>=0.4.22
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0

# LLM & Agents
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0

# Document Processing
pymupdf>=1.23.0

# Database (for checkpoints)
aiosqlite>=0.19.0

# Streaming & HTTP
httpx>=0.26.0
sse-starlette>=1.8.0
websockets>=12.0

# File Monitoring
watchdog>=4.0.0

# Logging
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
